+++ core/dbt/contracts/graph/components.py
@@ -4,6 +4,8 @@ from dbt.artifacts.resources import (
     FreshnessThreshold as FreshnessThresholdResource,
     Time as TimeResource,
 )
+from dbt_common.contracts.util import Replaceable
+from dbt_common.dataclass_schema import dbtClassMixin
 from typing import Dict, Optional
 
 
@@ -39,3 +41,27 @@ class FreshnessThreshold(FreshnessThresholdResource):
 
     def __bool__(self):
         return bool(self.warn_after) or bool(self.error_after)
+
+
+# Metrics, exposures,
+@dataclass
+class HasRelationMetadata(dbtClassMixin, Replaceable):
+    database: Optional[str]
+    schema: str
+
+    # Can't set database to None like it ought to be
+    # because it messes up the subclasses and default parameters
+    # so hack it here
+    @classmethod
+    def __pre_deserialize__(cls, data):
+        data = super().__pre_deserialize__(data)
+        if "database" not in data:
+            data["database"] = None
+        return data
+
+    @property
+    def quoting_dict(self) -> Dict[str, bool]:
+        if hasattr(self, "quoting"):
+            return self.quoting.to_dict(omit_none=True)
+        else:
+            return {}
